Split WS clients listeners module in three
