+++ src/server/WSClientListenerObserver.js
@@ -21,32 +21,18 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-const express = require('express');
-const path = require('path');
-const WSTaskServer = require('./WSTaskServer');
-const WebSocketPool = require('./WebSocketPool');
-const WSClientListenerExecutor = require('./WSClientListenerExecutor')
-const WSClientListenerObserver = require('./WSClientListenerObserver')
+const WSClientListener = require('./WSClientListener');
 
-const PORT = process.env.PORT || 3000;
+/**
+ * This client is able only to receive messages
+ * as a passive observer.
+ * Attempts to send a message end up with connection close.
+ */
+class WSClientListenerObserver extends WSClientListener {
+  onMessage(message) {
+    console.log(`Unexpected message from observer '${message}'`);
+    this.socket.close();
+  }
+}
 
-const STATIC_PATH = path.join(__dirname, '..', '..', 'dist');
-const INDEX_FILE = path.join(STATIC_PATH, 'index.html');
-
-const server = express()
-  .use(express.static(STATIC_PATH))
-  .use((req, res) => res.sendFile(INDEX_FILE))
-  .listen(PORT, () => {
-    console.log(`Pattern recognition server is listening on port ${PORT}`);
-  });
-
-const socketPool = new WebSocketPool(5);
-
-/* eslint-disable-next-line no-unused-vars */
-const wss = new WSTaskServer({
-  Executor: WSClientListenerExecutor,
-  Observer: WSClientListenerObserver,
-  server,
-  socketPool,
-  taskPath: '/',
-});
+module.exports = WSClientListenerObserver;
